# public/swagger.yaml
components:
    examples: {}
    headers: {}
    parameters: {}
    requestBodies: {}
    responses: {}
    schemas:
        Branch:
            properties:
                id: {}
                branch_name:
                    type: string
                branch_location:
                    type: string
                branch_city:
                    type: string
                branch_phone_number:
                    type: string
                manager_id:
                    type: number
                    format: double
            required:
                - id
                - branch_name
                - branch_location
                - branch_city
                - branch_phone_number
                - manager_id
            type: object
            additionalProperties: false
        BranchPayload:
            description: "\"branch_name\": \"BOIB121\"\r\n\"branch_loation\": \"near khandwa road\"\r\n\"branch_city\": \"indore\"\r\n\"branch_phone_number\": \"9770902921\"\r\n\"manager_id\": 121112"
            properties:
                branch_name:
                    type: string
                branch_location:
                    type: string
                branch_city:
                    type: string
                branch_phone_number:
                    type: string
                manager_id:
                    type: string
            required:
                - branch_name
                - branch_location
                - branch_city
                - branch_phone_number
                - manager_id
            type: object
            additionalProperties: false
        Accounts:
            properties:
                id:
                    type: number
                    format: double
                account_number:
                    type: string
                branch_id:
                    type: string
                customer_id:
                    type: string
                account_type:
                    type: string
                account_status:
                    type: string
            required:
                - id
                - account_number
                - branch_id
                - customer_id
                - account_type
                - account_status
            type: object
            additionalProperties: false
        Customer:
            properties:
                id:
                    type: number
                    format: double
                name:
                    type: string
                address:
                    type: string
                phone:
                    type: string
                email:
                    type: string
                password:
                    type: string
                branch_id:
                    type: string
                account_number:
                    type: string
                Update_At:
                    type: string
                    format: date-time
                account_id:
                    $ref: '#/components/schemas/Accounts'
                transection_id:
                    items:
                        $ref: '#/components/schemas/Transection'
                    type: array
            required:
                - id
                - name
                - address
                - phone
                - email
                - branch_id
                - account_number
                - Update_At
                - account_id
                - transection_id
            type: object
            additionalProperties: false
        Transection:
            properties:
                id:
                    type: number
                    format: double
                account_number:
                    type: string
                createdAt:
                    type: string
                payment_type:
                    type: string
                loan_id:
                    type: string
                transection_type:
                    type: string
                customer_id:
                    $ref: '#/components/schemas/Customer'
            required:
                - id
                - account_number
                - createdAt
                - payment_type
                - loan_id
                - transection_type
                - customer_id
            type: object
            additionalProperties: false
        ICustomer:
            properties:
                user:
                    $ref: '#/components/schemas/Customer'
                token:
                    type: string
            required:
                - user
                - token
            type: object
            additionalProperties: false
        UserRequest:
            description: "name: \"sachin Patel\"\r\naddress: \"new Rani Bangh Indore\"\r\nphone: \"9770902921\"\r\nemail: \"patelsachinsp269@gmail.com\"\r\npassword: \"sach123\"\r\nbranch_id: \"branch1211\"\r\n Update_At: 2022/5/23"
            properties:
                name:
                    type: string
                address:
                    type: string
                phone:
                    type: string
                email:
                    type: string
                password:
                    type: string
                branch_id:
                    type: string
                Update_At: {}
            required:
                - name
                - address
                - phone
                - email
                - password
                - branch_id
                - Update_At
            type: object
            additionalProperties: false
        LoginPayload:
            description: "email: \"patelsachinsp269@gmail.com\"\r\npassword: \"sach123\""
            properties:
                email:
                    type: string
                password:
                    type: string
            required:
                - email
                - password
            type: object
            additionalProperties: false
        IUser:
            properties:
                id:
                    type: number
                    format: double
            required:
                - id
            type: object
            additionalProperties: false
        Department:
            properties:
                department_id: {}
                department_name:
                    type: string
                head_of_department:
                    type: string
                discription:
                    type: string
            required:
                - department_id
                - department_name
                - head_of_department
                - discription
            type: object
            additionalProperties: false
        Employees:
            properties:
                id:
                    type: number
                    format: double
                name:
                    type: string
                address:
                    type: string
                phone:
                    type: string
                email:
                    type: string
                password:
                    type: string
                branch_id:
                    type: string
                account_number:
                    type: string
                Update_At:
                    type: string
                    format: date-time
                department:
                    type: string
                department_id:
                    type: string
            required:
                - id
                - name
                - address
                - phone
                - email
                - branch_id
                - account_number
                - Update_At
                - department
                - department_id
            type: object
            additionalProperties: false
        IEmployee:
            properties:
                user:
                    $ref: '#/components/schemas/Employees'
                token:
                    type: string
            required:
                - user
                - token
            type: object
            additionalProperties: false
        EmployeePayload:
            description: "name: \"sachin Patel\"\r\naddress: \"new Rani Bangh Indore\"\r\nphone: \"9770902921\"\r\nemail: \"patelsachinsp269@gmail.com\"\r\npassword: \"sach123\"\r\nbranch_id: \"branch1211\"\r\ndepartment:\"developer\"\r\ndepartment_id: \"12ddd2d\""
            properties:
                name:
                    type: string
                address:
                    type: string
                phone:
                    type: string
                email:
                    type: string
                password:
                    type: string
                branch_id:
                    type: string
                department:
                    type: string
                department_id:
                    type: string
            required:
                - name
                - address
                - phone
                - email
                - password
                - branch_id
                - department
                - department_id
            type: object
            additionalProperties: false
        AccountPayload:
            properties:
                account_number:
                    type: string
                branch_id:
                    type: string
                customer_id:
                    type: string
                account_type:
                    type: string
                account_statu:
                    type: string
            required:
                - account_number
                - branch_id
                - customer_id
                - account_type
                - account_statu
            type: object
            additionalProperties: false
    securitySchemes:
        authorization:
            type: apiKey
            name: authorization
            in: header
info:
    title: Bank-Management
    version: 0.0.1
    description: 'Awesome project developed with TypeORM.'
    contact: {}
openapi: 3.0.0
paths:
    /bank-branch/add-branch:
        post:
            operationId: AddBranch
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Branch'
            tags:
                - Branch
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BranchPayload'
    /customer/signup:
        post:
            operationId: Signup
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ICustomer'
            tags:
                - Customer
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserRequest'
    /customer/login:
        post:
            operationId: Login
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ICustomer'
            tags:
                - Customer
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginPayload'
    /customer/get-user-details:
        post:
            operationId: GetUserDetails
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Customer'
            tags:
                - Customer
            security:
                -
                    Authontication: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IUser'
    /customer/transection:
        post:
            operationId: GetUserTransection
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    anyOf: [{items: {$ref: '#/components/schemas/Customer'}, type: array}, {type: number, format: double}]
                                type: array
            tags:
                - Customer
            security:
                -
                    Authontication: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IUser'
    /bank-department:
        post:
            operationId: AddDepartment
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Department'
            tags:
                - Department
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            additionalProperties: false
                            type: object
    /employee/signup:
        post:
            operationId: Signup
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IEmployee'
            tags:
                - Employee
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/EmployeePayload'
    /employee/login:
        post:
            operationId: Login
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IEmployee'
            tags:
                - Employee
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginPayload'
    /employee/add-account:
        post:
            operationId: AddAccount
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Accounts'
            tags:
                - Employee
            security:
                -
                    authorization: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AccountPayload'
servers:
    -
        url: /